synth-data-pipeline/pyproject.toml
[project]
name = "synth-data-pipeline"
version = "0.1.0"
description = "Synthetic data generation pipeline for NanoChat fine-tuning"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "logfire>=4.14.2",
    "pydantic-ai>=1.6.0",
    "python-dotenv>=1.1.1",
    "textprompts>=1.0",
    "tqdm>=4.66.0",
    "openai>=1.0.0",
    "numpy>=1.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/synth_data_pipeline"]
